requirements.txt
Django==1.11.11
djangorestframework==3.12.4
